pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = "A comprehensive Python project template with API, CLI, CRUD, WebSocket, SSE, and REST client examples."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.110.0",
    "typer>=0.12.0",
    "httpx>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.0",
    "python-dotenv>=1.0.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=1.3.0",
    "respx>=0.23.1",
]

[project.scripts]
python-template = "python_template.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "PL", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["PLR2004", "S101"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
